Add files via upload
Visualizes data of overall and specialized ratings of flight features for British Airways. Each flight feature is rated out of a number of 5-- where 1 is least satisfactory and 5 is excellent. The ratings include averages of: cabin service, entertainment, food, ground service, seat comfort, value for money and overall rating. Among these ratings, the dashboard presents filters for: region, type of aircraft, seat type, and date in order to understand what variables contribute the most to the collected data. Furthermore, the visual aids presented in the dashboard isolate the key filters that may or may not contribute to the rating results. The line chart showcases the varying results of ratings per each feature between a timeline of March 2016 and October 2023. The map outlines continents and equips shading as a visual tool of comparison between results of every continent. The bar graph follows a similar line of reasoning between the top 9 aircrafts in conjunction to the various other aircrafts grouped as one. The color scheme allows a coherent and visually pleasing reading of the dashboard. The dashboard also includes the number of reviews per category and filter to accurately account for the significance each configuration of filters harbors.